Revert "Fixed typo for AuthentificationException, finalized the authentificat…"
