--- src/Main.java
@@ -1,2 +1,0 @@
-public class Main {
-}
