+++ src/auth/core/Admin.java
@@ -1,9 +1,4 @@
 package auth.core;
 
-/**
- * Admin is a superuser of this system. It has access to every action in this banking system.
- * @see User
- *
- */
-public class Admin extends User {
+public class Admin extends Role{
 }
