+++ src/auth/core/Teller.java
@@ -1,7 +1,4 @@
 package auth.core;
 
-/**
- * Teller is a user that can view anyone's account.
- */
-public class Teller extends User {
+public class Teller extends Role{
 }
