+++ src/auth/exceptions/AuthentificationException.java
@@ -1,5 +1,5 @@
 package auth.exceptions;
 
-public class TimeOutException extends RuntimeException{
+public class AuthentificationException extends RuntimeException {
     // TODO: Implement attributes and message display (toString) for this exception.
 }
