--- src/auth/exceptions/DuplicateAccountException.java
@@ -1,10 +1,0 @@
-package auth.exceptions;
-
-public class DuplicateAccountException extends RuntimeException {
-    private final String ERROR_MESSAGE = "Duplicate!";
-
-    @Override
-    public String getMessage() {
-        return ERROR_MESSAGE;
-    }
-}
