--- src/auth/exceptions/InsufficientFundsException.java
@@ -1,10 +1,0 @@
-package auth.exceptions;
-
-public class InsufficientFundsException extends RuntimeException{
-    private final String ERROR_MESSAGE = "Error! Wrong email or password.";
-
-    @Override
-    public String getMessage() {
-        return ERROR_MESSAGE;
-    }
-}
