+++ src/auth/exceptions/PasswordFormatException.java
@@ -1,0 +1,4 @@
+package auth.exceptions;
+
+public class PasswordFormatException extends RuntimeException{
+}
