+++ src/example/DivisionAction.java
@@ -1,7 +1,7 @@
 package example;
 
-import auth.core.User;
-import auth.exceptions.InvalidAuthenticationException;
+import auth.core.Role;
+import auth.exceptions.AuthentificationException;
 import auth.exceptions.TimeOutException;
 import bank.Account;
 import core.Action;
@@ -12,6 +12,8 @@ public class DivisionAction extends Action {
     private int quotient;
     private int remainder;
 
+    private boolean isAuthorized;
+
     @Override
     public void prepare() throws DivisionByZeroException, TimeOutException {
         // Division by zero!
@@ -19,7 +21,7 @@ public class DivisionAction extends Action {
             throw new RuntimeException();
         }
 
-        if (!isAuthorized()) {
+        if (!isAuthorized) {
             // Sometimes the connection times out, thus revisit the authentification of the user
             throw new TimeOutException();
         }
@@ -33,13 +35,18 @@ public class DivisionAction extends Action {
     }
 
     @Override
-    public void authorize(User user, Account account) throws InvalidAuthenticationException {
+    public void authorize(Role role, Account account) throws AuthentificationException {
         /*
            Authentify user with an account in the database
            If valid, set isAuthorized to true
            If invalid, throw AuthentificationException
          */
-        throw new InvalidAuthenticationException();
+        throw new AuthentificationException();
+    }
+
+    @Override
+    public boolean isAuthorized() {
+        return false;
     }
 
     public double getDividend() {
