+++ src/example/MockMain.java
@@ -1,7 +1,7 @@
 package example;
 
-import auth.core.User;
-import auth.exceptions.InvalidAuthenticationException;
+import auth.core.Role;
+import auth.exceptions.AuthentificationException;
 import auth.exceptions.TimeOutException;
 import auth.core.Customer;
 
@@ -23,9 +23,9 @@ public class MockMain {
         mockAction.setDividend(10);
         mockAction.setDivisor(5);
         try {
-            mockAction.authorize(new User() {
-            }, new ExampleAccount(new Customer("jack.sparrow@carribean.sea.com","12345","Jack","Sparrow")));
-        } catch (InvalidAuthenticationException e) {
+            mockAction.authorize(new Role() {
+            }, new ExampleAccount(new Customer("Jack")));
+        } catch (AuthentificationException e) {
             // Display bad authentification error, redirect flow of execution based on the activity diagram or use case diagram
             displayErrorGUI(e);
         }
